+++ netbox_branching/tables/columns.py
@@ -1,4 +1,8 @@
 import django_tables2 as tables
+from django.utils.translation import gettext_lazy as _
+
+from core.tables import ObjectChangeTable
+from utilities.tables import register_table_column
 
 __all__ = (
     'ConflictsColumn',
@@ -73,3 +77,12 @@ class DiffColumn(tables.TemplateColumn):
 
     def value(self, value):
         return str(value) if value else None
+
+
+# Register a table column to include the Branch associated with each record in ObjectChangeTable
+branch_column = tables.Column(
+    accessor=tables.A('application__branch'),
+    linkify=True,
+    verbose_name=_('Branch')
+)
+register_table_column(branch_column, 'branch', ObjectChangeTable)